+++ apps/daemon/src/index.ts
@@ -2,14 +2,13 @@ import "./tui/index.tsx";
 import "./log4js.ts";
 import { serve } from "bun";
 import { getLogger } from "log4js";
-import { fromThrowable } from "neverthrow";
 import appConfig from "./config";
 import applicationConfig from "./config";
 import { HonoApplication } from "./hono/HonoApplication.ts";
 import { recentLoggingEvent$ } from "./log4js.ts";
 import { ProdApplication } from "./ProdApplication.ts";
 import { startTui } from "./tui";
-import { openUrl } from "./utils/openUrl.ts";
+import { WebviewWorker } from "./webview";
 
 const APP_URL = "https://dcs-dropzone-container.flying-dice.workers.dev";
 
@@ -44,6 +43,22 @@ const bunServer = serve({
 
 logger.info(`🚀 Server running at ${bunServer.url}`);
 
+const webviewWorker: WebviewWorker = new WebviewWorker();
+
+async function handleGracefulShutdown() {
+	logger.info("Graceful shutdown initiated...");
+	logger.debug("Terminating webview worker...");
+	webviewWorker.terminate(APP_URL);
+
+	logger.debug("Stopping Bun server...");
+	await bunServer.stop();
+
+	logger.debug("Closing application...");
+	app.close();
+
+	logger.info("Shutdown complete.");
+}
+
 if (applicationConfig.app.tui_enabled) {
 	logger.info("Starting TUI...");
 	await startTui({
@@ -52,17 +67,32 @@ if (applicationConfig.app.tui_enabled) {
 		onEnableRelease: (releaseId) => app.enableRelease(releaseId),
 		onDisableRelease: (releaseId) => app.disableRelease(releaseId),
 		onRemoveRelease: (releaseId) => app.disableRelease(releaseId),
-		onOpenBrowser: () => openUrl(APP_URL),
+		onOpenBrowser: () => {
+			webviewWorker
+				.open(APP_URL, {
+					debug: process.env.NODE_ENV !== "production",
+				})
+				.match(
+					() => {},
+					(e) => {
+						throw new Error(e);
+					},
+				);
+		},
 		onQuit: async () => {
 			logger.info("Quit signal received from TUI, shutting down...");
-			await bunServer.stop(true);
-			process.exit(0);
+			await handleGracefulShutdown();
 		},
 	});
 }
 
 logger.debug("Bootstrap complete!");
 
-process.on("exit", (code) => {
-	logger.info(`Process exiting with code: ${code}`);
+process.on("SIGINT", async () => {
+	logger.info("SIGINT received, shutting down...");
+	await handleGracefulShutdown();
+});
+
+process.on("exit", async (code) => {
+	logger.info(`Process exiting with code ${code}`);
 });